Updated to Arduino 1.5.x+ library specifications.

+++ examples/EventControllerTester/EventControllerTester.ino
@@ -1,13 +1,13 @@
-#include "Arduino.h"
-#include "Streaming.h"
-#include "Array.h"
-#include "TimerOne.h"
-#include "Functor.h"
-#include "FunctorCallbacks.h"
-#include "EventController.h"
+#include <Arduino.h>
+#include <Streaming.h>
+#include <Array.h>
+#include <TimerOne.h>
+#include <Functor.h>
+#include <FunctorCallbacks.h>
+#include <EventController.h>
 
 
-const int BAUDRATE = 115200;
+const long BAUDRATE = 115200;
 const int LOOP_DELAY = 1000;
 const int CLOCK_PERIOD = 1000;
 const int CLOCK_START_DELAY = 2000;
